fix unit tests

+++ src/main/java/com/epam/ta/reportportal/job/SaveBinaryDataJob.java
@@ -131,7 +131,9 @@ public class SaveBinaryDataJob implements Runnable {
         } catch (IOException e) {
             LOGGER.error("Unable to save binary data", e);
         } finally {
-            ((CommonsMultipartFile) file).getFileItem().delete();
+            if (file instanceof CommonsMultipartFile) {
+                ((CommonsMultipartFile) file).getFileItem().delete();
+            }
         }
     }
